Move unschedulablePods struct to a separate file
